feat: use viper for cli command :rocket:
